Added global frame functionality
Now there is a global frame for the serialRobot class that is specified by the member Tbase (affine transformation matrix from the base frame to the global frame). Added new function members to serialRobot that return the unitTwists and transformation matrices in global coordinates.
Re-wrote the transformation matrix functions so that they are only function members of serialRobot (rather than being able to call transformMatDH(...) with a bunch of DH parameters).
Made the magSerialRobot class return magnet and magnetPos members in global coordinates, and updated the actuation matrix to be returned in global coordinates as well. This assumes that the global coordinates are the coil coordinates.