Added min SSV, WIP optimization
Added the m_calc_min_ssv() function member to the MagSerialRobot class. This function returns the minimum smallest singular value of the current actuation matrix (Mu) at the present robot configuration. Also, I attempted to write a code that automatically optimizes the robot magnetic parameters. This optimization code is still a work in progress. I probably should've made a separate branch. Oh well. I've been trying to figure out how to pass the robot parameters to the LMFunctor struct, because the optimization depends on the kinematics and dynamics of the MagSerialRobot.